+++ src/components/DuplicantMenu/components/MegaMenuItem/index.ts
@@ -1,0 +1,4 @@
+import MegaMenuItem from "./component";
+import connectMegaMenuItem from "./connector";
+
+export default connectMegaMenuItem(MegaMenuItem);